[FE] refactor: 프로젝트 디렉터리 구조를 컨벤션에 맞게 리팩터링한다.  (#861)
* refactor: 컨벤션에 맞게 수정

* feat: 고객모드 폴더 리팩터링

* refactor: 컨벤션에 맞게 디렉터리 경로 수정

* fix: 잘못된 DialPad import 변경

---------
