+++ frontend/src/pages/Admin/CustomCouponDesign/components/StampCustomModal/style.tsx
