+++ frontend/src/pages/Admin/ManageCafe/index.tsx
@@ -2,7 +2,7 @@ import Text from '../../../components/Text';
 import Button from '../../../components/Button';
 import { ManageCafeForm, ManageCafeGridContainer, PreviewContainer, Wrapper } from './style';
 import { useMemo } from 'react';
-import { PreviewImageWrapper } from '../CouponDesign/CustomCouponDesign/style';
+import { PreviewImageWrapper } from '../CustomCouponDesign/style';
 import LoadingSpinner from '../../../components/LoadingSpinner';
 import { useRedirectRegisterPage } from '../../../hooks/useRedirectRegisterPage';
 import { Cafe } from '../../../types/domain/cafe';
