+++ frontend/src/pages/Admin/ModifyCouponPolicy/components/CreatedType/style.tsx
