--- frontend/src/pages/Admin/ModifyCouponPolicy/logic/index.ts
@@ -1,13 +1,0 @@
-import ROUTER_PATH from '../../../../constants/routerPath';
-import { CouponCreated } from '../../../../types/domain/coupon';
-
-export const selectRoutePathByCreatedType = (createdType: CouponCreated) => {
-  switch (createdType) {
-    case 'template':
-      return ROUTER_PATH.templateCouponDesign;
-    case 'custom':
-      return ROUTER_PATH.customCouponDesign;
-    default:
-      throw new Error('[ERROR]: 쿠폰 생성 타입이 잘못되었습니다.');
-  }
-};
